indexdepartment screen rendering with backend data
